pyproject.toml
[project]
name = "git-cai-cli"
dynamic = ["version"]
description = "Use LLM to create git commit messages"
authors = [{ name = "Thorsten Foltz", email = "thorsten.foltz@live.com" }]
keywords = ["Git", "LLM", "Commit", "AI", "GenAI"]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "openai>=1.0",
    "google-genai>=1.39",
    "typer>=0.19.2",
    "requests>=2.32.5",
    "anthropic>=0.75.0",
    "groq>=0.37.0",
]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Version Control",
]

[project.urls]
Homepage = "https://github.com/thorstenfoltz/cai"
Issues = "https://github.com/thorstenfoltz/cai/issues"

[project.scripts]
git-cai = "git_cai_cli.cli:app"

[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"

[dependency-groups]
dev = [
    "build>=1.3.0",
    "pytest>=8.4.2",
    "twine>=6.2.0",
]
